move each route into it's own folder which also contains doc specific files
