+++ src/routes/auth/auth-doc.ts
@@ -1,0 +1,29 @@
+/*
+ * @api [post] /login
+ *  summary: "create session cookie"
+ *  description: "This endpoint uses firebase to create cookies for the admin to access privliged endpoints."
+ *  tags:
+ *    - Test Endpoints
+ *  produces:
+ *    - application/json
+ *  responses:
+ *    200:
+ *      description: creates session cookie.
+ *    401:
+ *      description: UNAUTHORIZED REQUEST!
+ *
+ */
+
+/*
+ * @api [get] /logout
+ *  summary: "clears session cookies"
+ *  description: "This endpoint clears session cookies, removing state history of user."
+ *  tags:
+ *    - Test Endpoints
+ *  produces:
+ *    - application/json
+ *  responses:
+ *    200:
+ *      description: redirect.
+ *
+ */
